# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/friends": {
      "get": {
        "operationId": "FriendsController_getFriendsList_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Tìm kiếm bạn bè theo tên hoặc email",
            "schema": {
              "example": "Nguyen Van",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bạn bè trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bạn bè thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedFriendsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách bạn bè của người dùng hiện tại",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/v1/friends/{friendId}": {
      "delete": {
        "operationId": "FriendsController_removeFriend_v1",
        "parameters": [
          {
            "name": "friendId",
            "required": true,
            "in": "path",
            "description": "ID của người bạn cần xóa",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa bạn bè thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ hoặc không phải bạn bè"
          },
          "401": {
            "description": "Không có quyền truy cập"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa bạn bè",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/v1/friends/{userId}/mutual": {
      "get": {
        "operationId": "FriendsController_getMutualFriends_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID của người dùng cần kiểm tra bạn bè chung",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bạn bè chung trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bạn bè chung thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutualFriendsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không có quyền truy cập"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách bạn bè chung với người dùng khác (người này có thể chưa phải là bạn bè)",
        "tags": [
          "Friends"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Swagger docs for the project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FriendResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID của người dùng",
            "example": "507f1f77bcf86cd799439011"
          },
          "fullName": {
            "type": "string",
            "description": "Họ và tên",
            "example": "Nguyen Van A"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar",
            "example": "https://example.com/avatar.jpg"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "nguyenvana@example.com"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "nguyenvana"
          },
          "bio": {
            "type": "string",
            "description": "Bio",
            "example": "I love sports!"
          },
          "favoritesSports": {
            "description": "Môn thể thao yêu thích",
            "example": [
              "football",
              "basketball"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "_id",
          "fullName",
          "email"
        ]
      },
      "PaginatedFriendsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Tổng số items",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "Trang hiện tại",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Số lượng items trên mỗi trang",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Tổng số trang",
            "example": 10
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Có trang tiếp theo không",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Có trang trước không",
            "example": false
          },
          "friends": {
            "description": "Danh sách bạn bè",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendResponseDto"
            }
          }
        },
        "required": [
          "page",
          "limit",
          "totalPages",
          "hasNextPage",
          "hasPrevPage",
          "friends"
        ]
      },
      "MutualFriendsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Tổng số items",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "Trang hiện tại",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Số lượng items trên mỗi trang",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Tổng số trang",
            "example": 10
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Có trang tiếp theo không",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Có trang trước không",
            "example": false
          },
          "currentUserId": {
            "type": "string",
            "description": "ID của người dùng hiện tại",
            "example": "507f1f77bcf86cd799439011"
          },
          "targetUserId": {
            "type": "string",
            "description": "ID của người dùng cần kiểm tra",
            "example": "507f1f77bcf86cd799439012"
          },
          "mutualFriends": {
            "description": "Danh sách bạn bè chung",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendResponseDto"
            }
          }
        },
        "required": [
          "page",
          "limit",
          "totalPages",
          "hasNextPage",
          "hasPrevPage",
          "currentUserId",
          "targetUserId",
          "mutualFriends"
        ]
      },
...
    }
  }
}